refactor NodeFeed and PostItem, get rid of TypeShell
